+++ demo/storage/big-data-test.ts
@@ -10,12 +10,13 @@ interface User extends DocumentObject {
   token: string;
 }
 
-const db = new AlwatrStorage<User>('junk-data', 'temp');
+const db = new AlwatrStorage<User>({name: 'junk-data', path: 'db'});
 
-db.ready.then(() => {
+db.readyPromise.then(() => {
   for (let i = 0; i < 10000; i++) {
     db.set({
       _id: random.string(4, 16),
+      _updatedBy: 'demo' + i,
       fname: random.string(4, 16),
       lname: random.string(4, 32),
       email: random.string(8, 32),