+++ packages/core/i18n/src/i18n.ts
@@ -16,9 +16,7 @@ export function initialI18n(options?: Partial<I18nOptions>): void {
     if (Object.prototype.hasOwnProperty.call(options, key)) {
       // configuration[key] = options[key];
       // WTF, mastmalize type.
-      (configuration[key as keyof I18nOptions] as unknown) = options[
-        key as keyof I18nOptions
-      ] as unknown;
+      (configuration[key as keyof I18nOptions] as unknown) = options[key as keyof I18nOptions] as unknown;
     }
   }