+++ packages/service/storage/src/index.ts
@@ -1,0 +1,6 @@
+import './route/home.js';
+import './route/update.js';
+import './route/get.js';
+import {logger} from './lib/config.js';
+
+logger.logOther('..:: Alwatr Storage Nanoservice API ::..');