+++ packages/starter/microservice/src/index.ts
@@ -1,5 +1,0 @@
-import './route/echo.js';
-import './route/home.js';
-import {logger} from './config.js';
-
-logger.logOther(`..:: Alwatr Microservice Starter Kit ::..`);