+++ packages/starter/nanoservice/src/index.ts
@@ -1,0 +1,5 @@
+import './route/home.js';
+import './route/echo.js';
+import {logger} from './lib/config.js';
+
+logger.logOther('..:: Alwatr Nanoservice Starter Kit ::..');