+++ packages/starter/nanoservice/src/lib/config.ts
@@ -1,0 +1,11 @@
+import {createLogger} from '@alwatr/logger';
+import {isNumber} from '@alwatr/math';
+
+export const config = {
+  /* eslint-disable @typescript-eslint/no-non-null-assertion */
+  port: isNumber(process.env.PORT) ? +process.env.PORT! : 80,
+};
+
+export const logger = createLogger('nanoservice-starter');
+
+logger.logProperty('config', config);